requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary  # For PostgreSQL
django-otp  # For OTP verification